requirements.txt
Jinja2~=3.1.2
pillow~=11.0.0
pyyaml~=6.0.2
rich~=12.5.1
